refactor useActionsDropdownHandlers
